Changed data type of credit earned to integer; Some code refactoring

+++ PlayerRecord.cs
@@ -5,13 +5,16 @@ using UnityEngine;
 public class PlayerRecord
 {
     public int PlayerID { get; set; }
+
     public string Name { get; set; }
-    public float CreditEarned { get; set; }
+
+    public int CreditEarned { get; set; }
+
     public long DateAchieved { get; set; }
 
-    public PlayerRecord(int ID = -1, string name, float credit, long dateAchieved)
+    public PlayerRecord(int ID = -1, string name, int credit, long dateAchieved)
     {
-        this.PlayerID=ID;
+        this.PlayerID = ID;
         this.Name = name;
         this.CreditEarned = credit;
         this.DateAchieved = dateAchieved;
